Added --vert flag, 0th stage of Translator setup
I've figured out how the initial formatting of the channeltypes work, and currently i've just made -Ch-> behave like a secure channel, and I've added an error message that displays if it finds this keyword without using --vert. Next step is to figure out what each of these 'Stages' the comments talk of are intended to do (the 0th one being formats and channels).