app rules working (except secret?)
Had to fix a bug relating to adding & x/=y facts, since there was still the issue of the . after the last rule, despite me thinking that I fixed that (was close though). In app5 the composed mac secret apply(pair(apply(mac,pair(apply(scrt,pair(C,s),N)),G)) is instead shown as simpy XmcscrtCs. Not sure if this is a problem, but I will ask during the meeting today. After this, app is only missing the attack state, which will likely be heavily hard-coded, but hopefully I can model it after app5 rule s.t. it's construction is based on the last rule of the specified app protocol.