separated App from Ch
The AnBMain.hs file now has a function that detects whether it's an App or Ch protocol, and passes along that boolean to relevant functions. In previous commits, I was working on the App protocol, but since the meeting showed that App wasn't fully converted to IF with certainty yet (my fault), I will now continue on Ch instead until the proof has been made *and* verified.